@@ -1,7 +1,7 @@
 {
   "name": "@expo/next-adapter",
-  "version": "4.0.0",
-  "description": "Adapter for using Next.js with Expo modules",
+  "version": "5.0.0",
+  "description": "Adapter for using Next.js to bundle Expo web projects",
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "scripts": {
@@ -12,18 +12,16 @@
     "lint": "eslint ."
   },
   "bin": {
-    "next-expo": "./build/cli/index.js"
+    "next-expo": "echo \"the next-expo command has been deprecated in favor of manual setup. See README for more.\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/expo/expo-cli.git",
+    "url": "git+https://github.com/expo/expo-cli.git",
     "directory": "packages/next-adapter"
   },
   "keywords": [
     "expo",
-    "expo-web",
-    "json",
-    "next",
+    "next.js",
     "react",
     "react-native",
     "react-dom",
@@ -35,35 +33,16 @@
   },
   "homepage": "https://github.com/expo/expo-cli/tree/main/packages/next-adapter#readme",
   "files": [
-    "babel",
-    "customize",
-    "template",
-    "build",
-    "document.js",
-    "document.d.ts"
+    "build"
   ],
-  "devDependencies": {
-    "@types/fs-extra": "^9.0.1",
-    "@types/node-fetch": "^2.5.8",
-    "@types/prompts": "^2.0.6"
-  },
   "peerDependencies": {
-    "next": "^11",
-    "react": "^17",
-    "react-native-web": "^0.17.0"
+    "expo": "^46",
+    "next": "^13",
+    "react": "^18",
+    "react-native-web": "~0.18.0",
+    "webpack": "^4.46.0 || ^5.74.0"
   },
   "publishConfig": {
     "access": "public"
-  },
-  "dependencies": {
-    "@expo/package-manager": "0.0.51",
-    "@expo/webpack-config": "0.16.20",
-    "babel-preset-expo": "^8.4.1",
-    "chalk": "^4.0.0",
-    "commander": "^4.0.1",
-    "fs-extra": "9.0.0",
-    "prompts": "^2.3.2",
-    "resolve-from": "^5.0.0",
-    "update-check": "^1.5.3"
   }
 }
