@@ -1,8 +1,6 @@
 #!/usr/bin/env ts-node-script
 import assert from 'assert';
 import program, { Command, Option } from 'commander';
-// @ts-ignore
-import replaceAll from 'string-replace-all-ponyfill';
 import stripAnsi from 'strip-ansi';
 
 import { registerCommands } from '../build/commands/index.js';
@@ -319,8 +317,8 @@ if (['markdown', 'md'].includes(process.argv[2])) {
     const booleanArg: FigArg = {
       name: 'boolean',
       suggestions: [
-        // { name: 'true', icon: ICON.true },
-        // { name: 'false', icon: ICON.false },
+        { name: 'true', icon: ICON.true },
+        { name: 'false', icon: ICON.false },
       ],
     };
     const helpOption: FigSpec = {
@@ -355,11 +353,6 @@ if (['markdown', 'md'].includes(process.argv[2])) {
       }
       return options;
     };
-
-    const getGeneratorArgs = (generator: string, args: Partial<FigArg> = {}): FigArg => ({
-      generators: `_gen[\`${generator}\`]`,
-      ...args,
-    });
 
     const stripTrailingDot = (str: string): string => str.replace(/\.$/, '');
     const capitalizeFirst = (str: string): string => str.charAt(0).toUpperCase() + str.slice(1);
@@ -404,10 +397,10 @@ if (['markdown', 'md'].includes(process.argv[2])) {
           Array.isArray(subcommand.args) && subcommand.args[0],
           'install [...packages] arg cannot be found'
         );
-        subcommand.args[0] = getGeneratorArgs('npm', {
+        subcommand.args[0] = {
           ...subcommand.args[0],
           debounce: true,
-        });
+        };
       }
 
       // Delete empty arrays for fig linting
@@ -430,7 +423,7 @@ if (['markdown', 'md'].includes(process.argv[2])) {
           .filter(Boolean) as string[];
         const [, argName] = option.flags.match(/(?:\[|<)([\w|-||]+)(?:\]|>)/) || [];
         const isRequired = argName ? option.flags?.includes('<') : false;
-        let args: FigArg = argName
+        const args: FigArg = argName
           ? {
               name: argName,
               isOptional: !isRequired,
@@ -470,19 +463,6 @@ if (['markdown', 'md'].includes(process.argv[2])) {
           args.name = 'platform';
         } else if (name[0] === '--config') {
           // config is deprecated so move it lower
-        } else if (name[0] === '--max-workers') {
-          args = getGeneratorArgs('max-workers', { ...args, name: 'Number of workers' });
-        }
-
-        if (command.name === 'run:ios') {
-          if (name.includes('--configuration')) {
-            args = getGeneratorArgs('xcode-configuration', args);
-          } else if (name.includes('--scheme')) {
-            args = getGeneratorArgs('xcode-scheme', args);
-            // User should choose schemes before devices as schemes can filter out devices
-          } else if (name.includes('--device')) {
-            args = getGeneratorArgs('xcode-devices', { ...args, isOptional: false });
-          }
         }
 
         const getOptionIcon = (option: OptionData) => {
@@ -612,87 +592,8 @@ if (['markdown', 'md'].includes(process.argv[2])) {
       }),
     };
 
-    // Replace all quotes around generators
-    const parsed = replaceAll(
-      JSON.stringify(figSpec, null, 2),
-      /"(_gen\[`.*`\])"/g,
-      (_: string, inner: string) => {
-        return inner;
-      }
-    );
-
-    const contents = `
-    // Copyright (c) 2021-present 650 Industries, Inc. (aka Expo)
-    // Generated by running \`yarn introspect fig\` in expo-cli/packages/expo-cli
-    // expo-cli@${require('../package.json').version}
-  
-    const _gen: Record<string, Fig.Generator> = {
-      "npm": {
-        script(context) {
-          if (context[context.length - 1] === "") return "";
-          const searchTerm = context[context.length - 1];
-          return \`curl -s -H "Accept: application/json" "https://api.npms.io/v2/search?q=$\{searchTerm}&size=20"\`;
-        },
-        postProcess(script: string) {
-          try {
-            const results: {
-              package: {
-                name: string;
-                description: string;
-              };
-              searchScore: number;
-            }[] = JSON.parse(script).results;
-            return results.map((item) => ({
-              name: item.package.name,
-              description: item.package.description,
-            })) as Fig.Suggestion[];
-          } catch (e) {
-            return [];
-          }
-        },
-      },
-      "xcode-configuration": {
-        script: "xcodebuild -project ios/*.xcodeproj -list -json",
-        postProcess: (script: string) => JSON.parse(script).project.configurations.map((name) => ({ name })),
-      },
-      "xcode-scheme": {
-        script: "xcodebuild -project ios/*.xcodeproj -list -json",
-        postProcess: (script: string) => JSON.parse(script).project.schemes.map((name) => ({ name })),
-      },
-      "xcode-devices": {
-        script: "xcrun xctrace list devices",
-        postProcess: (script: string) =>
-          script
-            .split("\\n")
-            .filter((item) => !item.match(/^=/))
-            .filter(Boolean)
-            .map((item) => item.split(/(.*?) (\\(([0-9.]+)\\) )?\\(([0-9A-F-]+)\\)/i))
-            // filter catalyst
-            .filter((item: string[] | null) => !!item?.[3])
-            .map(([, name, , osVersion, udid]) => ({
-              displayName: \`$\{name.trim()} ($\{osVersion})\`,
-              name: udid,
-            })),
-      },
-      "max-workers": {
-        script: "sysctl -n hw.ncpu",
-        postProcess: (script: string) => {
-          const count = Number(script);
-          return Array.from({ length: count }, (_, i) => {
-            const v = count - i;
-            return {
-              priority: v,
-              name: String(v),
-            };
-          });
-        }
-      }
-    }
-  
-    const completionSpec: Fig.Spec = ${parsed};
-
-    export default completionSpec;
-    `;
+    const contents = `const completionSpec: Fig.Spec = ${JSON.stringify(figSpec, null, 2)};
+export default completionSpec;`;
 
     // Generate a schema for https://github.com/withfig/autocomplete
 
