@@ -7,6 +7,7 @@ import { LoadingPageHandler, Project, UnifiedAnalytics, UrlUtils, Versions } fro
 import StatusEventEmitter from '../../analytics/StatusEventEmitter';
 import getDevClientProperties from '../../analytics/getDevClientProperties';
 import Log from '../../log';
+import { warnAboutLocalCLI } from '../../utils/migration';
 import { assertProjectHasExpoExtensionFilesAsync } from '../utils/deprecatedExtensionWarnings';
 import { profileMethod } from '../utils/profileMethod';
 import * as sendTo from '../utils/sendTo';
@@ -16,10 +17,11 @@ import { validateDependenciesVersionsAsync } from '../utils/validateDependencies
 import { ensureWebSupportSetupAsync } from '../utils/web/ensureWebSetup';
 import * as TerminalUI from './TerminalUI';
 import { installCustomExitHook, installExitHooks } from './installExitHooks';
-import { tryOpeningDevToolsAsync } from './openDevTools';
 import { NormalizedOptions, parseStartOptions } from './parseStartOptions';
 
 export async function actionAsync(projectRoot: string, options: NormalizedOptions): Promise<void> {
+  warnAboutLocalCLI(projectRoot, { localCmd: 'start' });
+
   Log.log(chalk.gray(`Starting project at ${projectRoot}`));
 
   // Add clean up hooks
@@ -55,12 +57,6 @@ export async function actionAsync(projectRoot: string, options: NormalizedOption
 
   // TODO: This is useless on mac, check if useless on win32
   const rootPath = path.resolve(projectRoot);
-
-  // Optionally open the developer tools UI.
-  await profileMethod(tryOpeningDevToolsAsync)(rootPath, {
-    exp,
-    options,
-  });
 
   if (Versions.gteSdkVersion(exp, '34.0.0')) {
     await profileMethod(ensureTypeScriptSetupAsync)(projectRoot);
