@@ -1,4 +1,5 @@
 import envinfo from 'envinfo';
+import { constants, promises } from 'fs';
 import path from 'path';
 
 import { findFile } from './helpers';
@@ -33,7 +34,20 @@ function getEnvironmentInfoAsync(): Promise<string> {
   );
 }
 
-export async function isBareWorkflowProject(projectRoot: string): Promise<boolean> {
+/* Poor mans implementation to prevent bloating the package size */
+export async function isInsideProjectAsync(projectRoot: string): Promise<boolean> {
+  try {
+    await promises.access(
+      path.join(projectRoot || process.cwd(), './package.json'),
+      constants.F_OK
+    );
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
+export async function isBareWorkflowProject(projectRoot: string): Promise<boolean | null> {
   const iosFound = await findFile(path.join(projectRoot, 'ios'), '.xcodeproj');
   const androidFound = await findFile(path.join(projectRoot, 'android'), '.gradle');
 
@@ -49,9 +63,11 @@ export async function actionAsync(projectRoot: string): Promise<void> {
   const info = await getEnvironmentInfoAsync();
   const lines = info.split('\n');
 
-  const workflow = (await isBareWorkflowProject(projectRoot)) ? 'bare' : 'managed';
-  lines.pop();
-  lines.push(`    Expo Workflow: ${workflow}`);
+  if (await isInsideProjectAsync(projectRoot)) {
+    const workflow = (await isBareWorkflowProject(projectRoot)) ? 'bare' : 'managed';
+    lines.pop();
+    lines.push(`    Expo Workflow: ${workflow}`);
+  }
 
   console.log(lines.join('\n') + '\n');
 }
