@@ -1,110 +1,17 @@
+import fs from 'fs';
+import path from 'path';
+
 import { setModulesMainApplication } from '../withAndroidModulesMainApplication';
 
+const fixturesPath = path.resolve(__dirname, 'fixtures');
+
 describe(setModulesMainApplication, () => {
-  it('should able to update from react native template', () => {
-    const rawContents = `
-package com.helloworld;
+  it('should able to update from react-native@>=0.71.0 template', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn071.java'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn071-updated.java'), 'utf8'),
+    ]);
 
-import android.app.Application;
-import com.facebook.react.PackageList;
-import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactNativeHost;
-import com.facebook.react.ReactPackage;
-import com.facebook.soloader.SoLoader;
-import java.util.List;
-
-public class MainApplication extends Application implements ReactApplication {
-
-  private final ReactNativeHost mReactNativeHost =
-      new ReactNativeHost(this) {
-        @Override
-        public boolean getUseDeveloperSupport() {
-          return BuildConfig.DEBUG;
-        }
-
-        @Override
-        protected List<ReactPackage> getPackages() {
-          @SuppressWarnings("UnnecessaryLocalVariable")
-          List<ReactPackage> packages = new PackageList(this).getPackages();
-          return packages;
-        }
-
-        @Override
-        protected String getJSMainModuleName() {
-          return "index";
-        }
-  };
-
-  @Override
-  public ReactNativeHost getReactNativeHost() {
-    return mReactNativeHost;
-  }
-
-  @Override
-  public void onCreate() {
-    super.onCreate();
-    SoLoader.init(this, /* native exopackage */ false);
-    // comment out initializeFlipper to keep test simpler
-    // initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
-  }
-}`;
-
-    const expectContents = `
-package com.helloworld;
-import android.content.res.Configuration;
-import expo.modules.ApplicationLifecycleDispatcher;
-import expo.modules.ReactNativeHostWrapper;
-
-import android.app.Application;
-import com.facebook.react.PackageList;
-import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactNativeHost;
-import com.facebook.react.ReactPackage;
-import com.facebook.soloader.SoLoader;
-import java.util.List;
-
-public class MainApplication extends Application implements ReactApplication {
-
-  private final ReactNativeHost mReactNativeHost =
-      new ReactNativeHostWrapper(this, new ReactNativeHost(this) {
-        @Override
-        public boolean getUseDeveloperSupport() {
-          return BuildConfig.DEBUG;
-        }
-
-        @Override
-        protected List<ReactPackage> getPackages() {
-          @SuppressWarnings("UnnecessaryLocalVariable")
-          List<ReactPackage> packages = new PackageList(this).getPackages();
-          return packages;
-        }
-
-        @Override
-        protected String getJSMainModuleName() {
-          return "index";
-        }
-  });
-
-  @Override
-  public ReactNativeHost getReactNativeHost() {
-    return mReactNativeHost;
-  }
-
-  @Override
-  public void onCreate() {
-    super.onCreate();
-    SoLoader.init(this, /* native exopackage */ false);
-    // comment out initializeFlipper to keep test simpler
-    // initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
-    ApplicationLifecycleDispatcher.onApplicationCreate(this);
-  }
-
-  @Override
-  public void onConfigurationChanged(Configuration newConfig) {
-    super.onConfigurationChanged(newConfig);
-    ApplicationLifecycleDispatcher.onConfigurationChanged(this, newConfig);
-  }
-}`;
     const contents = setModulesMainApplication(rawContents, 'java');
     expect(contents).toEqual(expectContents);
     // Try it twice...
@@ -112,94 +19,38 @@ public class MainApplication extends Application implements ReactApplication {
     expect(nextContents).toEqual(expectContents);
   });
 
-  it('should support another manually modified kotlin version MainApplication', () => {
-    const rawContents = `
-package com.helloworld
+  it('should able to update from react-native@>=0.68.0 template', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn068.java'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn068-updated.java'), 'utf8'),
+    ]);
 
-import android.app.Application
-import android.content.res.Configuration;
-import com.facebook.react.PackageList
-import com.facebook.react.ReactApplication
-import com.facebook.react.ReactNativeHost
-import com.facebook.react.ReactPackage
-import com.facebook.soloader.SoLoader
+    const contents = setModulesMainApplication(rawContents, 'java');
+    expect(contents).toEqual(expectContents);
+    // Try it twice...
+    const nextContents = setModulesMainApplication(contents, 'java');
+    expect(nextContents).toEqual(expectContents);
+  });
 
-class MainApplication : Application(), ReactApplication {
-  private val mReactNativeHost: ReactNativeHost = object : ReactNativeHost(this) {
-    override fun getUseDeveloperSupport(): Boolean {
-      return BuildConfig.DEBUG
-    }
+  it('should able to update from react-native@<0.68.0 template', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn064.java'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn064-updated.java'), 'utf8'),
+    ]);
 
-    override fun getPackages(): List<ReactPackage> {
-      return PackageList(this).packages
-    }
+    const contents = setModulesMainApplication(rawContents, 'java');
+    expect(contents).toEqual(expectContents);
+    // Try it twice...
+    const nextContents = setModulesMainApplication(contents, 'java');
+    expect(nextContents).toEqual(expectContents);
+  });
 
-    override fun getJSMainModuleName(): String {
-      return "index"
-    }
-  }
+  it('should support another manually modified kotlin version MainApplication', async () => {
+    const [rawContents, expectContents] = await Promise.all([
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn064.kt'), 'utf8'),
+      fs.promises.readFile(path.join(fixturesPath, 'MainApplication-rn064-updated.kt'), 'utf8'),
+    ]);
 
-  override fun getReactNativeHost(): ReactNativeHost {
-    return mReactNativeHost
-  }
-
-  override fun onCreate() {
-    super.onCreate()
-    SoLoader.init(this,  /* native exopackage */ false)
-    // comment out initializeFlipper to keep test simpler
-    // initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
-  }
-
-  override fun onConfigurationChanged(newConfig: Configuration) {
-    super.onConfigurationChanged(newConfig)
-  }
-}`;
-
-    const expectContents = `
-package com.helloworld
-import expo.modules.ApplicationLifecycleDispatcher
-import expo.modules.ReactNativeHostWrapper
-
-import android.app.Application
-import android.content.res.Configuration;
-import com.facebook.react.PackageList
-import com.facebook.react.ReactApplication
-import com.facebook.react.ReactNativeHost
-import com.facebook.react.ReactPackage
-import com.facebook.soloader.SoLoader
-
-class MainApplication : Application(), ReactApplication {
-  private val mReactNativeHost: ReactNativeHost = ReactNativeHostWrapper(this, object : ReactNativeHost(this) {
-    override fun getUseDeveloperSupport(): Boolean {
-      return BuildConfig.DEBUG
-    }
-
-    override fun getPackages(): List<ReactPackage> {
-      return PackageList(this).packages
-    }
-
-    override fun getJSMainModuleName(): String {
-      return "index"
-    }
-  })
-
-  override fun getReactNativeHost(): ReactNativeHost {
-    return mReactNativeHost
-  }
-
-  override fun onCreate() {
-    super.onCreate()
-    SoLoader.init(this,  /* native exopackage */ false)
-    // comment out initializeFlipper to keep test simpler
-    // initializeFlipper(this, getReactNativeHost().getReactInstanceManager());
-    ApplicationLifecycleDispatcher.onApplicationCreate(this)
-  }
-
-  override fun onConfigurationChanged(newConfig: Configuration) {
-    super.onConfigurationChanged(newConfig)
-    ApplicationLifecycleDispatcher.onConfigurationChanged(this, newConfig)
-  }
-}`;
     const contents = setModulesMainApplication(rawContents, 'kt');
     expect(contents).toEqual(expectContents);
     // Try it twice...
