@@ -1,12 +1,11 @@
-import {
-  ConfigPlugin,
-  withDangerousMod,
-  withXcodeProject,
-  XcodeProject,
-} from '@expo/config-plugins';
+import { ConfigPlugin, withDangerousMod } from '@expo/config-plugins';
 import fs from 'fs';
 import path from 'path';
+import resolveFrom from 'resolve-from';
 import semver from 'semver';
+import { ISA, XCBuildConfiguration } from 'xcparse';
+
+import { withXCParseXcodeProject, XCParseXcodeProject } from './withXCParseXcodeProject';
 
 type IosDeploymentTargetConfigPlugin = ConfigPlugin<{ deploymentTarget: string }>;
 
@@ -22,7 +21,11 @@ const withIosDeploymentTargetPodfile: IosDeploymentTargetConfigPlugin = (config,
     async config => {
       const podfile = path.join(config.modRequest.platformProjectRoot, 'Podfile');
       let contents = await fs.promises.readFile(podfile, 'utf8');
-      contents = updateDeploymentTargetPodfile(contents, props.deploymentTarget);
+      contents = await updateDeploymentTargetPodfile(
+        config.modRequest.projectRoot,
+        contents,
+        props.deploymentTarget
+      );
 
       await fs.promises.writeFile(podfile, contents);
       return config;
@@ -30,20 +33,86 @@ const withIosDeploymentTargetPodfile: IosDeploymentTargetConfigPlugin = (config,
   ]);
 };
 
-export function updateDeploymentTargetPodfile(contents: string, deploymentTarget: string): string {
-  return contents.replace(
-    /^(\s*platform :ios, ['"])([\d.]+)(['"])/gm,
-    (match, prefix, version, suffix) => {
-      if (semver.lt(toSemVer(version), toSemVer(deploymentTarget))) {
-        return `${prefix}${deploymentTarget}${suffix}`;
-      }
-      return match;
+// Because regexp //g is stateful, to use it multiple times, we should create a new one.
+function createPodfilePlatformRegExp() {
+  return /^(\s*platform :ios,\s*)(['"][\d.]+['"]|min_ios_version_supported)/gm;
+}
+
+async function parseVersionAsync(projectRoot: string, versionPart: string): Promise<string | null> {
+  let version;
+  if (versionPart === 'min_ios_version_supported') {
+    version = await lookupReactNativeMinIosVersionSupported(projectRoot);
+  } else {
+    version = versionPart.replace(/'"/g, '');
+  }
+  return version;
+}
+
+export async function updateDeploymentTargetPodfile(
+  projectRoot: string,
+  contents: string,
+  deploymentTarget: string
+): Promise<string> {
+  const matchResult = createPodfilePlatformRegExp().exec(contents);
+  if (matchResult) {
+    const version = await parseVersionAsync(projectRoot, matchResult[2]);
+    if (version && semver.lt(toSemVer(version), toSemVer(deploymentTarget))) {
+      return contents.replace(createPodfilePlatformRegExp(), (match, prefix, versionPart) => {
+        return `${prefix}'${deploymentTarget}'`;
+      });
     }
+  }
+  return contents;
+}
+
+export async function shouldUpdateDeployTargetPodfileAsync(
+  projectRoot: string,
+  targetVersion: string
+) {
+  const podfilePath = path.join(projectRoot, 'ios', 'Podfile');
+  const podfile = await fs.promises.readFile(podfilePath, 'utf-8');
+  const matchResult = createPodfilePlatformRegExp().exec(podfile);
+  if (!matchResult) {
+    console.warn(
+      'Unrecognized `ios/Podfile` content, will skip the process to update minimum iOS supported version.'
+    );
+    return false;
+  }
+
+  const version = await parseVersionAsync(projectRoot, matchResult[2]);
+  if (!version) {
+    console.warn(
+      'Unrecognized `ios/Podfile` content, will skip the process to update minimum iOS supported version.'
+    );
+    return false;
+  }
+
+  return semver.lt(toSemVer(version), toSemVer(targetVersion));
+}
+
+export async function lookupReactNativeMinIosVersionSupported(
+  projectRoot: string
+): Promise<string | null> {
+  const reactNativePodsScript = resolveFrom.silent(
+    projectRoot,
+    'react-native/scripts/react_native_pods.rb'
   );
+  if (!reactNativePodsScript) {
+    return null;
+  }
+  try {
+    const content = await fs.promises.readFile(reactNativePodsScript, 'utf-8');
+    const matchRepExp = /^def min_ios_version_supported\n\s*return\s*['"]([\d.]+)['"]/gm;
+    const matchResult = matchRepExp.exec(content);
+    if (matchResult) {
+      return matchResult[1];
+    }
+  } catch {}
+  return null;
 }
 
 const withIosDeploymentTargetXcodeProject: IosDeploymentTargetConfigPlugin = (config, props) => {
-  return withXcodeProject(config, config => {
+  return withXCParseXcodeProject(config, config => {
     config.modResults = updateDeploymentTargetXcodeProject(
       config.modResults,
       props.deploymentTarget
@@ -53,17 +122,19 @@ const withIosDeploymentTargetXcodeProject: IosDeploymentTargetConfigPlugin = (co
 };
 
 export function updateDeploymentTargetXcodeProject(
-  project: XcodeProject,
+  project: XCParseXcodeProject,
   deploymentTarget: string
-): XcodeProject {
-  const configurations = project.pbxXCBuildConfigurationSection();
-  for (const { buildSettings } of Object.values(configurations ?? {})) {
-    const currDeploymentTarget = buildSettings?.IPHONEOS_DEPLOYMENT_TARGET;
-    if (
-      currDeploymentTarget &&
-      semver.lt(toSemVer(currDeploymentTarget), toSemVer(deploymentTarget))
-    ) {
-      buildSettings.IPHONEOS_DEPLOYMENT_TARGET = deploymentTarget;
+): XCParseXcodeProject {
+  for (const section of Object.values(project.objects ?? {})) {
+    if (section.isa === ISA.XCBuildConfiguration) {
+      const { buildSettings } = section as XCBuildConfiguration;
+      const currDeploymentTarget = buildSettings?.IPHONEOS_DEPLOYMENT_TARGET;
+      if (
+        currDeploymentTarget &&
+        semver.lt(toSemVer(currDeploymentTarget), toSemVer(deploymentTarget))
+      ) {
+        buildSettings.IPHONEOS_DEPLOYMENT_TARGET = deploymentTarget;
+      }
     }
   }
   return project;
