@@ -1,9 +1,4 @@
-import { ConfigPlugin, withAppDelegate, withDangerousMod } from '@expo/config-plugins';
-import {
-  getAppDelegateObjcHeaderFilePath,
-  getPBXProjectPath,
-} from '@expo/config-plugins/build/ios/Paths';
-import { getDesignatedSwiftBridgingHeaderFileReference } from '@expo/config-plugins/build/ios/Swift';
+import { ConfigPlugin, IOSConfig, withAppDelegate, withDangerousMod } from '@expo/config-plugins';
 import {
   addObjcImports,
   insertContentsInsideObjcFunctionBlock,
@@ -12,14 +7,17 @@ import {
 import fs from 'fs';
 import { sync as globSync } from 'glob';
 import semver from 'semver';
-import xcode from 'xcode';
+
+import {
+  getDesignatedSwiftBridgingHeaderFileReference,
+  withXCParseXcodeProject,
+} from './withXCParseXcodeProject';
 
 export const withIosModulesAppDelegate: ConfigPlugin = config => {
   return withAppDelegate(config, config => {
-    config.modResults.contents =
-      config.modResults.language === 'objc'
-        ? updateModulesAppDelegateObjcImpl(config.modResults.contents, config.sdkVersion)
-        : updateModulesAppDelegateSwift(config.modResults.contents, config.sdkVersion);
+    config.modResults.contents = ['objc', 'objcpp'].includes(config.modResults.language)
+      ? updateModulesAppDelegateObjcImpl(config.modResults.contents, config.sdkVersion)
+      : updateModulesAppDelegateSwift(config.modResults.contents, config.sdkVersion);
     return config;
   });
 };
@@ -29,7 +27,7 @@ export const withIosModulesAppDelegateObjcHeader: ConfigPlugin = config => {
     'ios',
     async config => {
       try {
-        const appDelegateObjcHeaderPath = getAppDelegateObjcHeaderFilePath(
+        const appDelegateObjcHeaderPath = IOSConfig.Paths.getAppDelegateObjcHeaderFilePath(
           config.modRequest.projectRoot
         );
         let contents = await fs.promises.readFile(appDelegateObjcHeaderPath, 'utf8');
@@ -42,38 +40,31 @@ export const withIosModulesAppDelegateObjcHeader: ConfigPlugin = config => {
 };
 
 export const withIosModulesSwiftBridgingHeader: ConfigPlugin = config => {
-  return withDangerousMod(config, [
-    'ios',
-    async config => {
-      const { projectRoot } = config.modRequest;
-      const projectFilePath = getPBXProjectPath(projectRoot);
-      const project = xcode.project(projectFilePath);
-      project.parseSync();
-
-      const bridgingHeaderFileName = getDesignatedSwiftBridgingHeaderFileReference({ project });
-      if (!bridgingHeaderFileName) {
-        return config;
-      }
-      const [bridgingHeaderFilePath] = globSync(
-        `ios/${bridgingHeaderFileName.replace(/['"]/g, '')}`,
-        {
-          absolute: true,
-          cwd: projectRoot,
-        }
-      );
-      if (!bridgingHeaderFilePath) {
-        return config;
-      }
-      let contents = await fs.promises.readFile(bridgingHeaderFilePath, 'utf8');
-
-      if (!contents.match(/^#import\s+<Expo\/Expo\.h>\s*$/m)) {
-        contents = addObjcImports(contents, ['<Expo/Expo.h>']);
-      }
-
-      await fs.promises.writeFile(bridgingHeaderFilePath, contents);
+  return withXCParseXcodeProject(config, async config => {
+    const bridgingHeaderFileName = getDesignatedSwiftBridgingHeaderFileReference(config.modResults);
+    if (!bridgingHeaderFileName) {
       return config;
-    },
-  ]);
+    }
+
+    const [bridgingHeaderFilePath] = globSync(
+      `ios/${bridgingHeaderFileName.replace(/['"]/g, '')}`,
+      {
+        absolute: true,
+        cwd: config.modRequest.projectRoot,
+      }
+    );
+    if (!bridgingHeaderFilePath) {
+      return config;
+    }
+    let contents = await fs.promises.readFile(bridgingHeaderFilePath, 'utf8');
+
+    if (!contents.match(/^#import\s+<Expo\/Expo\.h>\s*$/m)) {
+      contents = addObjcImports(contents, ['<Expo/Expo.h>']);
+    }
+
+    await fs.promises.writeFile(bridgingHeaderFilePath, contents);
+    return config;
+  });
 };
 
 export function updateModulesAppDelegateObjcImpl(
@@ -102,6 +93,10 @@ export function updateModulesAppDelegateObjcImpl(
       /\[\[RCTRootView alloc\] initWithBridge:/g,
       '[self.reactDelegate createRootViewWithBridge:'
     );
+    contents = contents.replace(/\bRCTAppSetupDefaultRootView\((.+)\)/g, (match, group) => {
+      const [bridge, moduleName, initProps] = group.split(',').map((s: string) => s.trim());
+      return `[self.reactDelegate createRootViewWithBridge:${bridge} moduleName:${moduleName} initialProperties:${initProps}]`;
+    });
     contents = contents.replace(
       /\[UIViewController new\]/g,
       '[self.reactDelegate createRootViewController]'
@@ -122,9 +117,14 @@ export function updateModulesAppDelegateObjcHeader(
 
   // Replace parent class if needed
   contents = contents.replace(
+    /^(\s*@interface\s+AppDelegate\s+:\s+)RCTAppDelegate$/m,
+    '$1EXAppDelegateWrapper'
+  ); // react-native@>=0.71.0
+
+  contents = contents.replace(
     /^(\s*@interface\s+AppDelegate\s+:\s+)UIResponder(\s+.+)$/m,
     '$1EXAppDelegateWrapper$2'
-  );
+  ); // react-native@<0.71.0
 
   return contents;
 }
