@@ -15,7 +15,7 @@ function createModuleVersionChecker(name: string, currentVersion: string) {
         deprecated: pkgJson.deprecated,
       };
     },
-    `${name}-updates.json`,
+    `${name}-${currentVersion}-updates.json`,
     24 * 60 * 60 * 1000 // one day
   );
 
